+++ ios/Classes/InterstitialAd/AdsterInterstitialAd.swift
@@ -38,6 +38,32 @@ class AdsterInterstitialAd : NSObject{
 }
 
 extension AdsterInterstitialAd: MediationAdDelegate {
+    func onCarouselBannerAdLoaded(carouselBannerAd: any AdsFramework.MediationCarouselBannerAd) {
+    }
+
+    func onNativeRewardAdLoaded(nativeRewardAd: any AdsFramework.MediationNativeRewardAd) {
+    }
+
+    func onCarouselNativeAdLoaded(carouselNativeAd: any AdsFramework.MediationCarouselNativeAd) {
+    }
+
+    func onAppOpenAdLoaded(appOpenAd: any AdsFramework.MediationAppOpenAd) {
+    }
+
+    func onAdRevenuePaid(revenue: Double, adUnitId: String, network: String, currency: String, precisionType: AdsFramework.PrecisionType) {
+        adClickChannel.invokeMethod(
+            "onAdRevenuePaid",
+            arguments: adsterRevenueArguments(
+                widgetId: widgetId,
+                revenue: revenue,
+                adUnitId: adUnitId,
+                network: network,
+                currency: currency,
+                precisionType: precisionType
+            )
+        )
+    }
+
     func onBannerAdLoaded(bannerAd: any AdsFramework.MediationBannerAd) {
 
     }